Initial project structure
Has a test replay called 'test-square.w3g' where I do the following:

* select the town hall
* summon a peon (no rally point)
* select the peon
* move him 4 times in a square
* end the game

The debug print out shows which actions are being run in the blocks and by
which player.

A basic `UnitManager` class will handle the creation of players,
association of their units, and then signal the unit actions
to the simulations.

First goal is to get a basic X/Y position map that a unit takes during a
game, without any simulation of the game rules yet.