Add more registation support.
For the scenario where a unit makes a building
but the player doesn't select that building before
moving onto something else - we end up with a possible
unit registration for the wrong building.

We instead wait to register building till we have a
'certain' event like a trained unit ability.

For now we support registering an alter after
training a hero.