Add known object id checks.
We now track the 'worker' and 'townhall'
object iDs once we are sure we know them.

This lets us 'guess' what kind of unit
was clicked on when using an ability
with a target object ID.

Notably this would usually mean a worker clicked
a tree or goldmine.  Likely will also mean
the same for clicking shops?

Nuetral detection should add to this list eventually.

Fix some bugs with possible assignments
from selections.

Add support for spawning new units into the
players world when using ability with no
target from a building.