Re-write the assignPossibleSelectGroup method
The step is meant to associate units in our
'possible select list' to units in our players
current selection.

If we do find a match in the itemId1-2 then we continue
on to look thru the players units to see if we can
find a unit with a matching itemId

We only have one known itemId / unit - since the group
selection could be mixed.  So in the end this will
only 'register' one unknown unit from the list ever,
but it should do so in a way that makes sense for the
unit matching.

Also fix some little bugs - like the wrong itemId names
for some worker types.