Lots of work for hero item usage and slot swaps.
We now register hero items whenever we have a situation
where we know what the item is, like if you move an item
and it's the only thing in your inventory - we can assign
the objectId1-2 and also apparently the targetX,targetY
is a marker for the item as well - so we register this as the
knownItemX and knownItemY

Through cascading situations we try to track item "swaps" - the game
only tells us which item slot was swapped with, not what the item
that was in the slot got swapped to.  So we need to infer what
happened in a bunch of situations.

There are still a few cases to figure out - and the potential for us
to assign the wrong reigstration for a unit is still probably possible.

But this does a pretty good job of figuring out most things.