Updates for tracking items and unknown items.
We try to register all dropped and unknown clicked
objects in the new World class.  Later on checking
too see if we can re-use any of these objectIds to make
some known objects out of them.

We want to try and keep the created units limited to actual
actions in game - to avoid things like tomes dropping
but heroes not acutally using them.

Still need to do more work in actually using the items /
consuming the items in order to help aidi tracking.