Major patch - includes new wc3v client viewer
Fixes a number of major bugs with changing unit selections,
tab switching units, and many movement bugs.

Adds a new `client` folder which contains a simple web app
made in pure JS which allows the loading and rendering of
wc3v files.

This client right now just has one map background - concealed hill.
Many things are till WIP about it - but the jist of the rendering
is there.

A few major things are left to implement in the wc3v parser itself,
but pro matches seem to be parsing with most data correct.

Readme updates a new video is planned for the first initial release
coming soon.