Lots of bug fixes for selection groups.
We now track unit group selects and deselects more
accurately by paying attention to the updates and which
block they occur in.

The game will issue a ChangeSelection back to back in the
action block when a user deselects a group to add another.

Because there are points in the game when we have a selection group
of all unregistered units and more than one unit in a group - the
engine can only take a best guess at hoping the selection we
got registration data about is the first in the group.

When we detect a 'same block' ChangeSelection that doesn't end up in
an empty group AND has some number of unknown selection units - we
assume they should have been cleared up either by a unreg summon destroy
or units being killed. The engine handles it by deselecting these units
from the group.

Units that are assigned to Hotkey groups will persist even if in game
they have died.  We end up in a scenario where the ChangeSelection command
doesn't fully clear a group of units because we inserted a dead one
into their selection group - which the game doesn't try to deselect.

Eventually we might try to re-use the objectIds assigned to this unit
when dealing with a more certain unit registration choice - we
can add a protection that 'unregisters' a unit which is detected to be
possibly dead.

---

Some of the next things to fix are tab selecting through multiple building
groups.