Add special handling for badly registered summons.
Occasionally we'll be in a situation where we can be sure
the unit we destroyed got something registered wrong.

We handle this case by keeping all destroyed units
in a list and checking against their objectIds1-2 pair.

In cash-vs-foggy we had a case where multiple Elementals
were out and spawned but only one could be fully registered.
It happened to be the wrong one we picked - and this only really
matters because the summmon timer destroyed the unit that got
the registration values.  When a group definately needed that
objectId1-2 pair we then can look up that we've destroyed it
and restore it to a potentially correct unregistered unit
of the same itemId.